refactor `boilerplates/drizzle`
