+++ boilerplates/drizzle/files/database/drizzleDb.ts
@@ -1,6 +1,5 @@
 import Database from "better-sqlite3";
 import { drizzle } from "drizzle-orm/better-sqlite3";
-import * as schema from "./schema";
 
 const sqlite = new Database("sqlite.db");
-export const db = drizzle(sqlite, { schema });
+export const drizzleDb = drizzle(sqlite);
