+++ boilerplates/drizzle/files/database/seed.ts
@@ -1,8 +1,9 @@
-import { db } from "./db";
-import { todoTable } from "./schema";
+import { drizzleDb } from "./drizzleDb";
+import { todoTable } from "./schema/todos";
 
 async function seed() {
-  db.insert(todoTable)
+  drizzleDb
+    .insert(todoTable)
     .values([{ text: "Buy milk" }, { text: "Buy strawberries" }])
     .run();
