+++ boilerplates/drizzle/files/drizzle.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
   dialect: "sqlite",
-  schema: "./database/schema.ts",
+  schema: "./database/schema/*",
   out: "./drizzle",
   dbCredentials: {
     url: "sqlite.db",
