+++ boilerplates/drizzle/package.json
@@ -35,11 +35,11 @@
     "./drizzle.config": {
       "types": "./dist/types/drizzle.config.d.ts"
     },
-    "./database/schema": {
-      "types": "./dist/types/database/schema.d.ts"
+    "./database/drizzleDb": {
+      "types": "./dist/types/database/drizzleDb.d.ts"
     },
-    "./database/db": {
-      "types": "./dist/types/database/db.d.ts"
+    "./database/schema/todos": {
+      "types": "./dist/types/database/schema/todos.d.ts"
     },
     "./database/seed": {
       "types": "./dist/types/database/seed.d.ts"
@@ -50,11 +50,11 @@
       "drizzle.config": [
         "./dist/types/drizzle.config.d.ts"
       ],
-      "database/schema": [
-        "./dist/types/database/schema.d.ts"
+      "database/drizzleDb": [
+        "./dist/types/database/drizzleDb.d.ts"
       ],
-      "database/db": [
-        "./dist/types/database/db.d.ts"
+      "database/schema/todos": [
+        "./dist/types/database/schema/todos.d.ts"
       ],
       "database/seed": [
         "./dist/types/database/seed.d.ts"
