+++ boilerplates/telefunc/files/pages/todo/TodoList.telefunc.ts
@@ -1,9 +1,9 @@
-import { db } from "@batijs/drizzle/database/db";
-import { todoTable } from "@batijs/drizzle/database/schema";
+import { drizzleDb } from "@batijs/drizzle/database/drizzleDb";
+import { todoTable } from "@batijs/drizzle/database/schema/todos";
 
 export async function onNewTodo({ text }: { text: string }) {
   if (BATI.has("drizzle")) {
-    await db.insert(todoTable).values({ text });
+    await drizzleDb.insert(todoTable).values({ text });
   } else {
     // This is where you'd persist the data
     console.log("Received new todo", { text });
