+++ boilerplates/trpc/files/trpc/server.ts
@@ -1,5 +1,5 @@
-import { db } from "@batijs/drizzle/database/db";
-import { todoTable } from "@batijs/drizzle/database/schema";
+import { drizzleDb } from "@batijs/drizzle/database/drizzleDb";
+import { todoTable } from "@batijs/drizzle/database/schema/todos";
 import { initTRPC } from "@trpc/server";
 
 /**
@@ -28,7 +28,7 @@ export const appRouter = router({
     })
     .mutation(async (opts) => {
       if (BATI.has("drizzle")) {
-        await db.insert(todoTable).values({ text: opts.input });
+        await drizzleDb.insert(todoTable).values({ text: opts.input });
       } else {
         // This is where you'd persist the data
         console.log("Received new todo", { text: opts.input });
